requirements.txt
pathlib==1.0.1; python_version < "3.4"
# Test requirements
spacy-nightly>=3.0.0a26,<3.0.0a30
pytest>=4.1.0
